# 18-SB-Rest-IRCTC-App/src/main/java/com/ashokit/request/PassengerRequest.java
package com.ashokit.request;

import javax.xml.bind.annotation.XmlRootElement;


@XmlRootElement
public class PassengerRequest {
	
	private String name;
	private String email; 
	private String from;
	private String to;
	private String jdate;
	private String trainNum;
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public String getEmail() {
		return email;
	}
	public void setEmail(String email) {
		this.email = email;
	}
	public String getFrom() {
		return from;
	}
	public void setFrom(String from) {
		this.from = from;
	}
	public String getTo() {
		return to;
	}
	public void setTo(String to) {
		this.to = to;
	}
	public String getJdate() {
		return jdate;
	}
	public void setJdate(String jdate) {
		this.jdate = jdate;
	}
	public String getTrainNum() {
		return trainNum;
	}
	public void setTrainNum(String trainNum) {
		this.trainNum = trainNum;
	}
	
	
}